Added feature from issue 39 - back button brings you to the last place you were (album view or playlist/track list) If you go Control->Last view (back button)->Control (back pressed) then it resets the last activity so that pressing back will close the app, so as not to cause an endless loop Going back to an album view correctly gets all info including album ID
Uses a static var in Status, so it should ideally be rewritten to not use a static var. If anyone has more info on how to use Session or Status in both TracksActivity and ControlActivity without losing the var, then let me know and I'll rewrite it. Currently, info written to ControlActivity.status from within TrackActivity gets lost when ControlActivity is brought back up since it looks like all of the activities call finish() when switching to something else.